requirements.txt
# LexConductor - Python Dependencies
# IBM Dev Day AI Demystified Hackathon 2026
# Team: AI Kings 👑

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# IBM watsonx Integration
ibm-watsonx-ai>=0.2.0
ibm-watsonx-orchestrate>=1.0.0

# IBM Cloud Services
ibm-cloud-sdk-core>=3.16.0
ibmcloudant>=0.7.0
ibm-cos-sdk>=2.13.0

# Environment & Configuration
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Async Support
asyncio>=3.4.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
hypothesis>=6.92.0

# Code Quality (Linting, Formatting, Type Checking)
black>=24.1.0
ruff>=0.1.15
mypy>=1.8.0

# Pre-commit Hooks
pre-commit>=3.6.0

# YAML Linting
yamllint>=1.33.0

# Security
detect-secrets>=1.4.0
